Some follow up changes
* Moved player safe loading to PlayerObject's fromJSON function
* Moved JSONSet and JSONMap to a new file
* Some random type improvements
* Removed sourceFiles from player interface
