+++ src/NetscriptFunctions.ts
@@ -1781,7 +1781,6 @@ export const ns: InternalAPI<NSFull> = {
       city: Player.city,
       location: Player.location,
       bitNodeN: Player.bitNodeN,
-      sourceFiles: new Map(Player.sourceFiles),
       totalPlaytime: Player.totalPlaytime,
       jobs: cloneDeep(Player.jobs),
       factions: Player.factions.slice(),
