+++ src/NetscriptFunctions/Formulas.ts
@@ -94,9 +94,8 @@ export function NetscriptFormulas(): InternalAPI<IFormulas> {
       numPeopleKilled: 0,
       money: 0,
       city: CityName.Sector12,
-      location: "",
+      location: LocationName.TravelAgency,
       bitNodeN: 0,
-      sourceFiles: new Map(),
       totalPlaytime: 0,
       jobs: {},
       factions: [],
