+++ src/ScriptEditor/NetscriptDefinitions.d.ts
@@ -51,7 +51,6 @@ interface Player extends Person {
   jobs: Record<string, string>;
   factions: string[];
   bitNodeN: number;
-  sourceFiles: Map<number, number>;
   totalPlaytime: number;
   location: string;
 }
